tempo indicator updated according to PDF; song list bugs fixed

+++ Backbeater/SoundLib/WindowQueue.h
@@ -15,7 +15,7 @@
 
 -(instancetype)initWithCapacity:(NSInteger)capacity;
 
--(instancetype)enqueue:(NSInteger)value;
+-(instancetype)enqueue:(Float64)value;
 
 -(void)clear;
